refactor rn list style, ref #560
