+++ components/list/index.md
@@ -32,10 +32,10 @@ source: design
 | arrow      | 箭头方向(右,上,下), 可选`horizontal`,`up`,`down`,`empty`，如果是`empty`则存在对应的dom,但是不显示   | String |   无  |
 | align    |    Flex 子元素垂直对齐，可选`top`,`middle`,`bottom`  | String   | `middle` |
 | onClick    | 点击事件的回调函数 | (): void |  无  |
-| error    | 报错样式,右侧文字颜色变成橙色 | Boolean  | `false`  |
+| error(`web only`)    | 报错样式,右侧文字颜色变成橙色 | Boolean  | `false`  |
 | multipleLine    | 多行 | Boolean  | `false`  |
 | wrap    | 是否换行，默认情况下，文字超长会被隐藏， | Boolean  | `false`  |
-| activeStyle(`Web only`)    | 自定义active的样式 | Object  |   |
+| activeStyle(`web only`)    | 自定义active的样式 | Object  |   |
 
 ### List.Item.Brief
